Update: Moved Components
Todo:
Create Sub Pages with the created Components
Make a fake Algorithm for Data
