+++ src/views/Analytics/compoments/graphs/GraphCard.vue
